Rett feil etter merge/flytting/rename
