+++ src/pages/Sykmeldingside/components/Infopanel/Infopanel.less
@@ -1,4 +1,4 @@
-@import '../../basic.less';
+@import '../../../../basic.less';
 
 .infopanel {
     background-color: white;
