+++ src/pages/Sykmeldingside/components/Infopanel/panelelementer/PrognoseSeksjon.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import tekster from '../infopanel-tekster';
+import tekster from '../Infopanel-tekster';
 import { Prognose } from '../../../../../types/sykmeldingTypes';
 import EtikettMedTekst from '../layout/EtikettMedTekst';
