+++ src/pages/Sykmeldingside/components/Statuspanel/Statuspanel.less
@@ -1,4 +1,4 @@
-@import '../../basic.less';
+@import '../../../../basic.less';
 
 .statuspanel {
     display: flex;
